+++ ui/options/OptionMetadata.lua
@@ -16,8 +16,6 @@ local optionDetails = {
             {value = "flat", label = Addon.L["OPT_BAR_STYLE_FLAT"]},
             {value = "vertical", label = Addon.L["OPT_BAR_STYLE_VERTICAL"]},
             {value = "circular", label = Addon.L["OPT_BAR_STYLE_CIRCULAR"]},
-            {value = "reputation", label = Addon.L["OPT_BAR_STYLE_REPUTATION"]},
-            {value = "segmented", label = Addon.L["OPT_BAR_STYLE_SEGMENTED"]}
         },
         commandKeys = {"style", "mode", "barstyle"}
     },